% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.R
\name{dropdown_lineplot}
\alias{dropdown_lineplot}
\title{A line plot with selectable y variables}
\usage{
dropdown_lineplot(df, x, yvars)
}
\arguments{
\item{df}{A data.frame}

\item{x}{The column name (as a string) of the x variable}

\item{yvars}{The column names (as strings) to be used as y variables}
}
\value{
A plotly object
}
\description{
A line plot with selectable y variables
}
\examples{
set.seed(42)
df <- lapply(1:26, function(x) rnorm(50))
df <- as.data.frame(df)
colnames(df) <- letters
df[["seq"]] <- 1:50
dropdown_lineplot(df, x="seq", yvars = letters)
}
