% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked_scatter_plot.R
\name{linked_scatter_plot}
\alias{linked_scatter_plot}
\title{linked_scatter_plot}
\usage{
linked_scatter_plot(
  x,
  y,
  links,
  groups = rep("1", length(x)),
  text = rep("", length(x)),
  title = "",
  xlab = "",
  xlim = range(x, na.rm = TRUE) * 1.1,
  ylim = range(y, na.rm = TRUE) * 1.1,
  ylab = "",
  colours = NULL
)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{links}{HTML links for each point}

\item{groups}{Groups to color the points by.}

\item{text}{Hovertext for the points}

\item{title}{Plot title}

\item{xlab, ylab}{X/Y axis title}

\item{xlim, ylim}{Range of X/Y axis}

\item{colours}{Colours to use for the groups}
}
\description{
A scatterplot with links
}
\examples{
linked_scatter_plot(
 1:10,
 1:10,
 rep("https://www.google.com", 10),
 groups = sample(letters[1:2], 10, replace=TRUE))
}
