% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{plotly_boxplot}
\alias{plotly_boxplot}
\title{A boxplot with jittered points}
\usage{
plotly_boxplot(x, box_hoverinfo = FALSE,
  colours = RColorBrewer::brewer.pal(ncol(x), "Paired"))
}
\arguments{
\item{x}{A data.frame}

\item{box_hoverinfo}{Show hoverinfo for the boxplot trace? (logical)}

\item{colours}{The colours to be used for the x-axis groups}
}
\value{
A plotly object
}
\description{
A boxplot with jittered points
}
\examples{
 plotly_boxplot(iris[, -5])
}
