% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{plotly_density}
\alias{plotly_density}
\title{Density plot using plotly}
\usage{
plotly_density(mat, title = "", xlab = "", palette = viridis)
}
\value{
A plotly htmlwidget
}
\description{
Creates a density plot of all columns of a matrix or data.frame like
structure.
}
\examples{
plotly_density(iris[, -5], title="Density plot of Iris dataset",
     xlab = "Value")

plotly_density(voomed_GBM$E, title="Density plot of voomed GBM data",
     xlab = "Expression level")
}
