% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyutils.R
\name{plotlyutils-pkg}
\alias{plotlyutils-pkg}
\title{The \pkg{plotlyutils} package}
\description{
Provides functions generating nice custom plotly plots as standalone
HTMLwidgets.
}
\author{
Alan O'Callaghan
}
