% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectable_histogram.R
\name{selectable_histogram}
\alias{selectable_histogram}
\title{selectable_histogram}
\usage{
selectable_histogram(
  vars,
  title = "",
  histnorm = c("probability density", "", "percent", "probability", "density")
)
}
\arguments{
\item{vars}{data.frame or matrix of variables
(each column will be a separate histogram).}

\item{title}{Plot title}

\item{histnorm}{See https://plot.ly/javascript/reference/#histogram-histnorm}
}
\description{
A scatterplot with dropdown menus to select the x/y variables and
the variable used to colour the points
}
\examples{
df <- lapply(1:10, function(...) rnorm(1000))
df <- as.data.frame(df)
colnames(df) <- paste("Histogram", letters[1:10])
selectable_histogram(df, "Interactive histogram")

}
