% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectable_scatter_plot.R
\name{selectable_scatter_plot}
\alias{selectable_scatter_plot}
\title{linked_scatterplot}
\usage{
selectable_scatter_plot(coords, colours, title = "")
}
\arguments{
\item{title}{Plot title}
}
\description{
A scatterplot with links
}
\examples{
pcs <- prcomp(mtcars)
colours <- mtcars[, c("cyl", "vs", "am", "gear", "carb")]
colours[] <- lapply(colours, as.character)
selectable_scatter_plot(
  pcs,
  colours,
  "interactive PCA plot"
)
}
