% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyutils.R
\docType{data}
\name{voomed_GBM}
\alias{voomed_GBM}
\title{Voom-transformed log2 counts per million for TCGA glioblastoma samples}
\source{
Derived from GBMdata
}
\description{
See data-raw for details.
}
\references{
https://genomebiology.biomedcentral.com/articles/10.1186/gb-2014-15-2-r29
}
\keyword{data}
