% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectable_scatter_plot.R
\name{selectable_scatter_plot}
\alias{selectable_scatter_plot}
\title{selectable_scatter_plot}
\usage{
selectable_scatter_plot(coords, colours, title = "")
}
\arguments{
\item{coords}{data.frame or matrix of point co-ordinates
Each column will be an entry in the X and Y drop-down menus}

\item{colours}{data.frame of variables used to colour points}

\item{title}{Plot title}
}
\description{
A scatterplot with dropdown menus to select the x/y variables and
the variable used to colour the points
}
\examples{
pcs <- prcomp(mtcars)
colours <- mtcars[, c("cyl", "vs", "am", "gear", "carb")]
colours[] <- lapply(colours, as.character)
selectable_scatter_plot(
  pcs$x,
  colours,
  "Interactive PCA plot"
)
}
